Major refactoring of Parser such that Parser is now an Interface with various Parser classes specific to each Command class implementing it. ParserManager then manages all the Parser classes.
